修正済み

+++ app/controllers/items_controller.rb
@@ -37,9 +37,12 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
+   if item = Item.find(params[:id])
     item.destroy
     redirect_to root_path
+   else
+    redirect_to item_path
+   end
   end
